backend/requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.8.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
distro==1.9.0
fastapi==0.115.8
filelock==3.17.0
fsspec==2025.2.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.28.1
idna==3.10
Jinja2==3.1.4
jiter==0.8.2
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numpy==2.2.2
openai==1.61.0
packaging==24.2
pillow==11.0.0
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.0.1
python-multipart==0.0.20
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.2
setuptools==70.2.0
sniffio==1.3.1
starlette==0.45.3
sympy==1.13.1
tokenizers==0.21.0
torch==2.6.0+cpu
torchaudio==2.6.0+cpu
torchvision==0.21.0+cpu
tqdm==4.67.1
transformers==4.48.2
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
